WIP - add day 4 star 1
Doing this in elm has been pretty painful in a lot of ways but also
pretty good practice for piping in particular.

I finally got to the point where I feel pretty sure that I've identified
the guard who sleeps the most (rendered on the page) and broken down
which minutes they spend asleep most often (logged to console) and
unfortunately there are 4 distinct minutes which each occur 15 times...

I need a single minute and I got 4 and I'm not sure how to proceed.

I've been poking at this for a few hours and so I'm going to stop now
and sleep since I'm up quite late (I'm in Finland) and have a lunch
reservation tomorrow and also I ought to wake up early enough to enjoy
the previous daylight hours.